+++ app/todos/components/todo.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default class Todo extends React.Component {
+  render () {
+    return <div>
+      { this.props.todo.text }
+    </div>
+  }
+}
